RRTMGP in its own directory (kernels don't move)
