+++ pyrte_rrtmgp/tests/test_data_validation.py
@@ -3,7 +3,7 @@ import netCDF4  # noqa
 import pytest
 import xarray as xr
 
-from pyrte_rrtmgp.rrtmgp_data_files import GasOpticsFiles
+from pyrte_rrtmgp.rrtmgp.data_files import GasOpticsFiles
 
 from pyrte_rrtmgp.examples import RFMIP_FILES
 from pyrte_rrtmgp.examples import load_example_file
